cleaned up main interface
