+++ src/main.rs
@@ -1,23 +1,14 @@
 
 extern crate wthrs;
-use wthrs::weather_opt::{WeatherOpt, Response};
-use wthrs::weather::Forecast;
-use tokio;
+use wthrs::weather_opt::WeatherOpt;
+// use tokio;
 use anyhow::Error;
 
 // use structopt::StructOpt;
 
 #[tokio::main]
 async fn main() -> Result<(), Error> {
-
-    let mut opt = WeatherOpt::from_args();
-    let resp: Response = opt.cmd.run().await?;
-
-    match resp {
-       Response::RequestResponse(forecast) => println!("{:}", forecast),
-       Response::ConfigSuccessResponse => println!("Success! Config saved.")
-    }
-
+    WeatherOpt::parse_args().await?;
     Ok(())
 
 }
